# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Novus Markets Trading Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/profile:
    patch:
      operationId: updateProfile
      tags: [users]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/kyc:
    get:
      operationId: getKycStatus
      tags: [users]
      summary: Get KYC status
      responses:
        "200":
          description: KYC status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KycStatus"
    post:
      operationId: submitKyc
      tags: [users]
      summary: Submit KYC documents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KycInput"
      responses:
        "200":
          description: KYC submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KycStatus"
...
  /users/change-password:
    post:
      operationId: changePassword
      tags: [users]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /users/2fa:
    patch:
      operationId: toggle2fa
      tags: [users]
      summary: Toggle 2FA
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Toggle2faInput"
      responses:
        "200":
          description: 2FA toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── Accounts ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ── Auth ────────────────────────────────────────────────────────────────
...
    User:
      type: object
      required: [id, email, firstName, lastName, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: ["string", "null"]
        preferredCurrency:
          type: string
          default: USD
        role:
          type: string
          enum: [user, admin]
        twoFaEnabled:
          type: boolean
        notificationsEnabled:
          type: boolean
        theme:
          type: string
          enum: [dark, light]
          default: dark
        createdAt:
          type: string
          format: date-time
    ProfileUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        preferredCurrency:
          type: string
        notificationsEnabled:
          type: boolean
        theme:
          type: string
          enum: [dark, light]
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 6
    Toggle2faInput:
      type: object
      required: [enabled]
      properties:
        enabled:
          type: boolean
    KycStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [unverified, pending, verified, rejected]
        documentType:
          type: ["string", "null"]
        submittedAt:
          type: ["string", "null"]
        reviewedAt:
          type: ["string", "null"]
        rejectionReason:
          type: ["string", "null"]
    KycInput:
      type: object
      required: [documentType, documentData]
      properties:
        documentType:
          type: string
          enum: [national_id, passport, drivers_license]
        documentData:
          type: string
          description: Base64 encoded document or document reference
...